+++ cmd/find.go
@@ -49,7 +49,7 @@ var findCmd = &cobra.Command{
 			core.OnErrorFail(errors.New("no Provider Selected"), "Selected Provider Not avilable (yet)")
 		}
 		log.Debug(string("Outputing List as " + options.Output + " Format"))
-		fmt.Println(provider.RunResult(p, options.Output))
+		fmt.Println(provider.PrintoutResults(p, options.Output))
 	},
 }