+++ cmd/struct.go
@@ -5,6 +5,9 @@ type FlagsOptions struct {
 	Output        string `json:"output,omitempty"`
 	Overwrite     bool   `json:"overwrite,omitempty"`
 	Backup        bool   `json:"backup,omitempty"`
+	Merge         bool   `json:"merge,omitempty"`
+	ForceMerge    bool   `json:"force_merge,omitempty"`
+	UiSize        int    `json:"uiSize,omitempty"`
 	DryRun        bool   `json:"dry-run,omitempty"`
 	AwsAuth       bool   `json:"aws_auth,omitempty"`
 	AwsAssumeRole bool   `json:"aws_assume_role,omitempty"`