+++ provider/structs.go
@@ -15,6 +15,9 @@ type CommandOptions struct {
 	Overwrite     bool
 	Combined      bool
 	Backup        bool
+	Merge         bool
+	ForceMerge    bool
+	UiSize        int
 	DryRun        bool
 	AwsAuth       bool
 	AwsRoleString string