+++ .github/workflows/CHANGELOG.md
@@ -1,7 +1,8 @@
 # k8f 
 ## Changes
+- Improved Connect description (`k8f connect -h`)
+- Added Merge Flag to combine new K8S and already existed K8S in kubeconfig file - inspired by the [kubecm](https://github.com/sunny0826/kubecm) by @sunny0826
 ## Bugfix
-- fixed az cli run and connect AKS
 ## Known Issue
 - connect command with "dry-run" flag will default to Json insted of Yaml
 <!-- ## Bugfix -->